requirements.txt
langchain_community
langchain_core
langchain-openai
python-dotenv
streamlit
langchain-openai
langchain
